requirements.txt
pydicom==2.3.0
requests==2.28.1
dicom2jpg==0.1.9
Pillow~=9.2.0